pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "patgen"
version = "0.1.0"
description = "Two-stage OpenAI pipeline: Event-B/Rust → NL brief → PAT/CSP-style output"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
patgen = "patgen.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["patgen*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
